+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - `{line}` value is now one-based.
 - All dirty files are now saved before running a command.
 - Added `{module}` variable which returns a dotted path to active file.
-- Added `justTesting.customExpressions` setting.
+- Added `justTesting.expressions` setting.
 
 ## 0.4.0